[type-stage] extract scope-generic machinery into scope-resolution.mjs
Pure refactor, zero behavior change: LocalScopeTypeSpecnion statics,
_BaseTypeSpecnion, PatchedTypeSpecnion and HierarchicalScopeTypeSpecnion
move to scope-resolution.mjs under scope-generic names
(LocalScopeProperties, _BaseScopeProperties, PatchedScopeProperties,
HierarchicalScopeProperties); mapSetProperties and
collectPropertyGeneratorEntries move alongside. Demarcation symbols stay
in type-specnion.mjs, which re-exports everything under the legacy names
- no consumer import changes. Foundation for the node-properties channel
(plan Phase 2).

model: moonshotai/kimi-k3
provider: openrouter
agent: goose v1.46.0
