+++ .prettierignore
@@ -38,3 +38,4 @@
 lib/assets/glyph-groups.yaml
 !lib/js/components/cascading-map.mjs
 !lib/js/components/cascading-map.test.mjs
+!lib/js/components/layouts/type-stage/scope-resolution.mjs
